Be more resilient against deleted posts
Previously, when replying or editing, the bridge queries properties of
the relevant post. However, such posts might have been deleted while the
bridge is not running, and, despite being present in the database, do
not actually exist.

This commit catches such errors and removes the offending posts from the
database.