Avoid using AppServiceBot
Again in order to eventually move away from matrix-appservice-bridge.

The main function that required replacement was isRemoteUser. This
simply checks the regex we put in the registration file. It was noticed
that the regex we used was overzealous - it claims matrix users from
other homeservers too. The commit fixes that as well.