Remove matrix-appservice-bridge dependency
We use matrix-js-sdk and matrix-appservice directly instead.

matrix-appservice-bridge comes with a lot of extra baggage that we do
not use.  For example, it includes a database (nedb), a logging library
(winston, as opposed to loglevel which comes with matrix-js-sdk) and
metrics (prom-client).

It also enforces a lot of dependency choice on us, e.g. json validator,
command line parser, etc.

Fixes #39