Leave default mattermost channels
Puppets now leave default mattermost channels if they are not in the
corresponding matrix room.

This requires working around a mattermost bug, where (I believe) joining
a team twice simultaneously causes one of them to return early when the
joining process has not completed. To avoid this, the client remembers
join team requests in progress and avoids sending a second request when
one is already pending.

Since I failed to reproduce this outside of the bridge, a mattermost
issue is not opened. This is in a sense an opposite problem to
https://github.com/mattermost/mattermost-server/issues/15366