Use exponential backoff when channel join 500's
Mattermost runs into race conditions when joining a channel twice
simultaneously:
https://github.com/mattermost/mattermost-server/issues/15366
In such situations, one of the join requests returns a 500 error but the
user eventually joins the channel anyway.

Previously, we dealt with this by checking if the user is in a channel
whenever we get at 500. If so, we counts it as success. However, this
still breaks in some edge cases, as evidenced in

https://github.com/dalcde/matrix-appservice-mattermost/runs/1083604332

I believe this happens because after we check that we are indeed in the
channel, it is possible in rare cases that we haven't actually finished
the joining process. If we perform actions assuming we have already
finished joining, this can cause breakages.

In this commit, we change to retrying the channel join with exponential
backoff. We then take a successful join to mean the joining has
completed. This assumption is not necessarily well-founded, as evidenced
by the previous commit ( ab3a38c5b9cb832e20eb36a7f0aca443b1909d7e ).
However, it should work better than what we have now.