Stop using bufferutil dependency
This is used by `ws` to make masking and unmasking more efficient. In
practice, this is only used when sending out messages in the websocket,
which is rare in our use cases.

Removing this dependency ensures the runtime dependencies are all pure
javascript, which makes it easier to package and distribute.